// frontend/src/types/contract.ts
export type Role =
  | 'super_admin'
  | 'admin'
  | 'manager'
  | 'academic_staff'
  | 'finance_staff'
  | 'society_rep'
  | 'student'
  | 'alumni'
  | 'past_alumni';

export type UserStatus = 'active' | 'suspended' | 'pending';
// ...
export type Faculty = 'FOC' | 'FOB' | 'FOE';
export type Programme = 'BSC-SE' | 'BSC-CS' | 'BBA' | 'BENG-CE';
export type YearGroup = 1 | 2 | 3 | 4;
// ...
export interface User {
  id: string;
  email: string;
  displayName: string;
  role: Role;
  status: UserStatus;
  faculty: Faculty | null;
  programme: Programme | null;
  yearGroup: YearGroup | null;
  studentId: string | null;
  phone: string | null;
  smsOptIn: boolean;
  societyIds: string[];
  department: string | null;
  alumniGradYear: number | null;
  fcmTokens: string[];
  notifPrefs: {
    push: boolean;
    categories: Record<string, boolean>;
  };
  lastLoginAt: string;
  createdAt: string;
}